Longifizied all snprintf format specifiers (because those functions are 16bit on AmigaOS).
git-svn-id: https://svn.aros.org/svn/aros/trunk@23014 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/libs/muimaster/classes/application.c
@@ -329,7 +329,7 @@ static ULONG Application_New(struct IClass *cl, Object *obj, struct opSet *msg)
 
 	for (i = 1; i < 1000; i++)
 	{
-	    snprintf(portname, 255, "%s.%d", data->app_Base, i);
+	    snprintf(portname, 255, "%s.%ld", data->app_Base, (LONG)i);
 	    if (!find_application_by_base(cl, obj, portname))
 		break;
 	}
